+++ Gemfile
@@ -1,0 +1,9 @@
+source 'https://rubygems.org'
+
+gem 'github-pages', group: :jekyll_plugins
+gem 'jekyll-octicons'
+
+gemspec
+
+# Added at 2017-08-08 13:15:27 -0700 by github:
+gem "jekyll-redirect-from", "~> 0.12.1"
