+++ _layouts/default.html
@@ -20,7 +20,7 @@
       ga('send', 'pageview');
     </script>
 
-    <script type="text/javascript" src="{{ site.baseurl }}/js/main.js"></script>
+    <script type="text/javascript" src="{{ site.baseurl }}/main.js"></script>
   </body>
 
 </html>
