+++ main.scss
@@ -1,3 +1,6 @@
+---
+---
+
 @import 'tools/tools';
 @import 'modules/buttons';
 @import 'modules/notif';
