+++ script/bootstrap
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+set -e
+export CC=gcc
+
+echo "==> Installing gem dependencies…"
+bundle check --path vendor/gems 2>&1 > /dev/null || {
+  time bundle install --binstubs bin --path vendor/gems
+}
