+++ script/cibuild
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+set -e
+
+DISABLE_WHITELIST=1 bundle exec jekyll build
+gem build github-website-starter.gemspec
